@@ -1,5 +1,5 @@
 module DynamicSearch
   module Version
-    VERSION = '0.4.3'
+    VERSION = '0.5.1'
   end
 end
