@@ -4,8 +4,9 @@ class TestResource < ActiveRecord::Base
   cattr_accessor :columns
   self.columns = []
 
-  def self.column(name, sql_type = nil, default = nil, null = true)
-    columns << ActiveRecord::ConnectionAdapters::Column.new(name.to_s, default, sql_type.to_s, null)
+  def self.column(name, sql_type = :string, default = nil, null = true)
+    type = ActiveModel::Type.lookup(sql_type)
+    columns << { name: name.to_s, type: type, default: default, null: null }
   end
 
   belongs_to :parent_resource, class_name: 'TestResource'
